Fix date-time regex for format validator
Previously the date-time regex required fields with a date-time format to end with the 'Z' offset indicating 0 offset from utc. However, json schema uses the rfc3339#section-5.6 standard which allows both the 'Z' and a time-numoffset format (e.g, -08:00, +02:00) for the utc offset. This was causing the validator to fail on date-time values that it should have succeeded. This change correctly validate either the 'Z' format or the time-numoffset for utc offsets in date-time format properties.